--- itms.client/src/app/user-details/comments/comments.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { CommentsComponent } from './comments.component';
-
-describe('CommentsComponent', () => {
-  let component: CommentsComponent;
-  let fixture: ComponentFixture<CommentsComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [CommentsComponent]
-    });
-    fixture = TestBed.createComponent(CommentsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
