+++ itms.client/src/app/user-details/laptop-details/laptop-details.component.html
@@ -47,7 +47,7 @@
       </div>
 
       
-        <app-comments [laptopDetails]="laptop" [userId]="userId"></app-comments>
+      <app-user-comments [laptopDetails]="laptop" [userId]="userId"></app-user-comments>
    
 
     </div>
