+++ itms.client/src/app/user-details/user-comments/user-comments.component.css
@@ -10,52 +10,54 @@
   border-radius: 0.25rem;
   background: #EAEEF6;
 }
+
   .hi.collapsed {
-    height: 90px; 
+    height: 90px;
   }
 
 .comment-container {
   max-height: 300px;
   overflow-y: auto;
   padding: 1rem;
-  
 }
 
-.comment-container.collapsed {
-    max-height: 60px; 
-}
-
-  .comment-container.collapsed .comment-header {
-    overflow: hidden;
+  .comment-container.collapsed {
+    max-height: 60px;
   }
 
-  .comment-container.collapsed .chevron-down {
-    margin-left: auto; 
-  }
+    .comment-container.collapsed .comment-header {
+      overflow: hidden;
+    }
+
+    .comment-container.collapsed .chevron-down {
+      margin-left: auto;
+    }
 
 .chevron {
   justify-content: flex-end;
   align-items: flex-start;
   display: flex;
   align-items: center;
-  cursor:pointer;
+  cursor: pointer;
 }
+
 .chevron-down {
   filter: invert(100%);
   margin-left: 35rem;
-  width:30px;
-  height:30px;
-  opacity:0.7;
-
+  width: 30px;
+  height: 30px;
+  opacity: 0.7;
 }
-.avatar{
-    width: 2rem;
+
+.avatar {
+  width: 2rem;
 }
 
 .comment-avatar {
   margin-top: 1rem;
   margin-left: 0%;
 }
+
 .comment-content {
   margin-top: 1rem;
   margin-left: 1%;
@@ -65,14 +67,14 @@
   height: 5% !important;
 }
 
-.user-details{
-    margin-left: 1.3%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
+.user-details {
+  margin-left: 1.3%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
 }
+
 .avatar-supercomponent {
-  
 }
 
 
@@ -89,8 +91,8 @@
   align-items: flex-start;
   gap: 1rem;
   align-self: stretch;
-  height:auto;
-  width:100%;
+  height: auto;
+  width: 100%;
 }
 
 .comment-box {
@@ -98,8 +100,8 @@
   padding: 1rem;
   align-items: flex-start;
   gap: 0.625rem;
-  height:auto;
-  width:100%;
+  height: auto;
+  width: 100%;
   align-self: stretch;
   border-radius: 0.25rem;
   background: #FFF;
