+++ itms.client/src/app/user-details/user-comments/user-comments.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { UserCommentsComponent } from './user-comments.component';
+
+describe('UserCommentsComponent', () => {
+  let component: UserCommentsComponent;
+  let fixture: ComponentFixture<UserCommentsComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      declarations: [UserCommentsComponent]
+    });
+    fixture = TestBed.createComponent(UserCommentsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
